+++ EPPZGeometry/EPPZGeometry_Polygon_Inspector.cs
@@ -1,0 +1,18 @@
+﻿using UnityEngine;
+using System.Collections;
+
+
+namespace EPPZGeometry
+{
+	
+	
+	public class EPPZGeometry_Polygon_Inspector : MonoBehaviour
+	{
+
+
+		public Polygon polygon;
+		public int currentEdgeIndex = 0;
+
+
+	}
+}